FIT9/requirements.txt
# Recommended: Use Python 3.10
streamlit==1.35.0
pandas==2.2.2
mysql-connector-python==8.3.0
plotly==5.22.0
altair==5.3.0
openai==1.30.1
mediapipe==0.10.11
opencv-python==4.9.0.80
